Generate the Map Icons at runtime
Generate the Map Icons at runtime on-demand. We may move this to an
external program at some point but for the moment this will do.

We are using the Spatial Illusions JavaScript library embedded into the
app as its API is very nice and easy to use compared to other C#
libraries