Add the ability to spawn a US NASAM site
Add the ability to spawn a US NASAM site, coalition will depend on what
country the US country belongs to.

This is not intended as a full feature but as a demonstrator of spawning
simple ground units.

For this to be usful we would probably need templates in YAML files that
the program can deserialize and send to DCS-gRPC so that people can
create their own templates.